requirements.txt
ib-insync
pandas
loguru
PyYAML
numpy

ta

pandas_ta
